Revert "refactor logctx_format_flush"
This reverts commit 43843fa088fd5b673c78979fefee1f75f3eca1f9.

Conflicts:

	logpool.c

+++ logpool.h
@@ -67,7 +67,7 @@ struct logCtx {
     logapi_t *formatter;
     logfmt_t logkey;
     logfmt_t fmt[LOGFMT_MAX_SIZE];
-    int logfmt_size;
+    long logfmt_size;
 };
 
 struct ltrace {
